ABuild test reorganization
