+++ projects/abuild/test/build_cache_test.cpp
@@ -1,8 +1,10 @@
-import atest;
 import abuild;
 import abuild_test_utilities;
 
+using atest::assert_;
+using atest::assert_fail;
 using atest::expect;
+using atest::expect_fail;
 using atest::suite;
 using atest::test;
