--- projects/abuild/test/sources_test.cpp
@@ -1,211 +1,0 @@
-import abuild;
-import abuild_test_utilities;
-
-using atest::assert_;
-using atest::assert_fail;
-using atest::expect;
-using atest::expect_fail;
-using atest::suite;
-using atest::test;
-
-static const auto testSuite = suite("abuild::Projects::sources()", [] {
-    test("root project", [] {
-        TestCache testCache;
-        TestProject testProject{"build_test_project_scanner",
-                                {"main.cpp",
-                                 "header.hpp"}};
-
-        abuild::BuildCache cache{testCache.file()};
-        abuild::ProjectScanner{cache, testProject.projectRoot()};
-
-        const std::string source = (testProject.projectRoot() / "main.cpp").string();
-
-        assert_(asVector(cache.sources()))
-            .toBe(std::vector<std::string>{
-                source});
-
-        expect(cache.sources()[source]["project"].GetString()).toBe(std::string{"build_test_project_scanner"});
-        expect(cache.sources()[source]["modified"].GetInt64()).toBe(lastModified(source));
-    });
-
-    test("root project with src dir", [] {
-        TestCache testCache;
-        TestProject testProject{"build_test_project_scanner",
-                                {"main.cpp",
-                                 "src/source.hpp",
-                                 "src/build_test_project_scanner/othersource.cpp"}};
-
-        abuild::BuildCache cache{testCache.file()};
-        abuild::ProjectScanner{cache, testProject.projectRoot()};
-
-        const std::string source1 = (testProject.projectRoot() / "main.cpp").string();
-        const std::string source2 = (testProject.projectRoot() / "src" / "build_test_project_scanner" / "othersource.cpp").string();
-
-        assert_(asVector(cache.sources()))
-            .toBe(std::vector<std::string>{
-                source1,
-                source2});
-
-        expect(cache.sources()[source1]["project"].GetString()).toBe(std::string{"build_test_project_scanner"});
-        expect(cache.sources()[source1]["modified"].GetInt64()).toBe(lastModified(source1));
-
-        expect(cache.sources()[source2]["project"].GetString()).toBe(std::string{"build_test_project_scanner"});
-        expect(cache.sources()[source2]["modified"].GetInt64()).toBe(lastModified(source2));
-    });
-
-    test("subprojects with sources", [] {
-        TestCache testCache;
-        TestProject testProject{"build_test_project_scanner",
-                                {"projects/myapp/src/myheader.cpp",
-                                 "projects/mylib/mylib.cc",
-                                 "projects/mylib/mysublib/projects/mysublib.cxx"}};
-
-        abuild::BuildCache cache{testCache.file()};
-        abuild::ProjectScanner{cache, testProject.projectRoot()};
-
-        const std::string source1 = (testProject.projectRoot() / "projects" / "myapp" / "src" / "myheader.cpp").string();
-        const std::string source2 = (testProject.projectRoot() / "projects" / "mylib" / "mylib.cc").string();
-        const std::string source3 = (testProject.projectRoot() / "projects" / "mylib" / "mysublib" / "projects" / "mysublib.cxx").string();
-
-        assert_(asVector(cache.sources()))
-            .toBe(std::vector<std::string>{
-                source1,
-                source2,
-                source3});
-
-        expect(cache.sources()[source1]["project"].GetString()).toBe(std::string{"myapp"});
-        expect(cache.sources()[source1]["modified"].GetInt64()).toBe(lastModified(source1));
-
-        expect(cache.sources()[source2]["project"].GetString()).toBe(std::string{"mylib"});
-        expect(cache.sources()[source2]["modified"].GetInt64()).toBe(lastModified(source2));
-
-        expect(cache.sources()[source3]["project"].GetString()).toBe(std::string{"mylib.mysublib"});
-        expect(cache.sources()[source3]["modified"].GetInt64()).toBe(lastModified(source3));
-    });
-
-    test("include external headers", [] {
-        TestCache testCache;
-        TestProjectWithContent testProject{"build_test_project_scanner",
-                                           {{"main.cpp", "#include <vector>\n#include <string>"}}};
-
-        abuild::BuildCache cache{testCache.file()};
-        abuild::ProjectScanner{cache, testProject.projectRoot()};
-        abuild::CodeScanner{cache};
-
-        const std::string source = (testProject.projectRoot() / "main.cpp").string();
-
-        assert_(asVector(cache.sources()))
-            .toBe(std::vector<std::string>{
-                source});
-
-        expect(asVector(cache.sources()[source]["includes_external"]))
-            .toBe(std::vector<std::string>{
-                "string",
-                "vector"});
-    });
-
-    test("include local headers", [] {
-        TestCache testCache;
-        TestProjectWithContent testProject{"build_test_project_scanner",
-                                           {{"main.cpp", "#include \"header.hpp\"\n#include \"other_header.hpp\""}}};
-
-        abuild::BuildCache cache{testCache.file()};
-        abuild::ProjectScanner{cache, testProject.projectRoot()};
-        abuild::CodeScanner{cache};
-
-        const std::string source = (testProject.projectRoot() / "main.cpp").string();
-
-        assert_(asVector(cache.sources()))
-            .toBe(std::vector<std::string>{
-                source});
-
-        expect(asVector(cache.sources()[source]["includes_local"]))
-            .toBe(std::vector<std::string>{
-                "header.hpp",
-                "other_header.hpp"});
-    });
-
-    test("import local headers", [] {
-        TestCache testCache;
-        TestProjectWithContent testProject{"build_test_project_scanner",
-                                           {{"main.cpp", "import \"header.hpp\";\nexport import \"other_header.hpp\";"}}};
-
-        abuild::BuildCache cache{testCache.file()};
-        abuild::ProjectScanner{cache, testProject.projectRoot()};
-        abuild::CodeScanner{cache};
-
-        const std::string source = (testProject.projectRoot() / "main.cpp").string();
-
-        assert_(asVector(cache.sources()))
-            .toBe(std::vector<std::string>{
-                source});
-
-        expect(asVector(cache.sources()[source]["import_includes_local"]))
-            .toBe(std::vector<std::string>{
-                "header.hpp",
-                "other_header.hpp"});
-    });
-
-    test("import external header", [] {
-        TestCache testCache;
-        TestProjectWithContent testProject{"build_test_project_scanner",
-                                           {{"main.cpp", "import <vector>;\nexport import <string>;"}}};
-
-        abuild::BuildCache cache{testCache.file()};
-        abuild::ProjectScanner{cache, testProject.projectRoot()};
-        abuild::CodeScanner{cache};
-
-        const std::string source = (testProject.projectRoot() / "main.cpp").string();
-
-        assert_(asVector(cache.sources()))
-            .toBe(std::vector<std::string>{
-                source});
-
-        expect(asVector(cache.sources()[source]["import_includes_external"]))
-            .toBe(std::vector<std::string>{
-                "string",
-                "vector"});
-    });
-
-    test("import modules", [] {
-        TestCache testCache;
-        TestProjectWithContent testProject{"build_test_project_scanner",
-                                           {{"main.cpp", "import mymodule;\nexport import othermodule;"}}};
-
-        abuild::BuildCache cache{testCache.file()};
-        abuild::ProjectScanner{cache, testProject.projectRoot()};
-        abuild::CodeScanner{cache};
-
-        const std::string source = (testProject.projectRoot() / "main.cpp").string();
-
-        assert_(asVector(cache.sources()))
-            .toBe(std::vector<std::string>{
-                source});
-
-        expect(asVector(cache.sources()[source]["import_modules"]))
-            .toBe(std::vector<std::string>{
-                "mymodule",
-                "othermodule"});
-    });
-
-    test("import module partition", [] {
-        TestCache testCache;
-        TestProjectWithContent testProject{"build_test_project_scanner",
-                                           {{"main.cpp", "import : mymodule;\nexport import : othermodule;"}}};
-
-        abuild::BuildCache cache{testCache.file()};
-        abuild::ProjectScanner{cache, testProject.projectRoot()};
-        abuild::CodeScanner{cache};
-
-        const std::string source = (testProject.projectRoot() / "main.cpp").string();
-
-        assert_(asVector(cache.sources()))
-            .toBe(std::vector<std::string>{
-                source});
-
-        expect(asVector(cache.sources()[source]["import_module_partitions"]))
-            .toBe(std::vector<std::string>{
-                "mymodule",
-                "othermodule"});
-    });
-});
